M-23 - Fix mqtt handler tests. (#28)
* M-23 - Fix mqtt handler tests.


* M-23 - Rename vars. Restore test case.


---------

+++ mqtt/handler.go
@@ -181,13 +181,18 @@ func (h *handler) Publish(c *session.Client, topic *string, payload *[]byte) {
 
 // Subscribe - after client successfully subscribed
 func (h *handler) Subscribe(c *session.Client, topics *[]string) {
-	sub, err := h.getSubcription(c, topics)
+	if c == nil {
+		h.logger.Error(LogErrFailedSubscribe + (ErrClientNotInitialized).Error())
+		return
+	}
+
+	subs, err := h.getSubcriptions(c, topics)
 	if err != nil {
 		h.logger.Error(LogErrFailedSubscribe + err.Error())
 		return
 	}
 
-	for _, s := range sub {
+	for _, s := range subs {
 		err = h.service.CreateSubscription(context.Background(), s)
 		if err != nil {
 			h.logger.Error(LogErrFailedSubscribe + (ErrSubscriptionAlreadyExists).Error())
@@ -198,13 +203,18 @@ func (h *handler) Subscribe(c *session.Client, topics *[]string) {
 
 // Unsubscribe - after client unsubscribed
 func (h *handler) Unsubscribe(c *session.Client, topics *[]string) {
-	sub, err := h.getSubcription(c, topics)
+	if c == nil {
+		h.logger.Error(LogErrFailedUnsubscribe + (ErrClientNotInitialized).Error())
+		return
+	}
+
+	subs, err := h.getSubcriptions(c, topics)
 	if err != nil {
 		h.logger.Error(LogErrFailedSubscribe + err.Error())
 		return
 	}
 
-	for _, s := range sub {
+	for _, s := range subs {
 		h.service.RemoveSubscription(context.Background(), s)
 		if c == nil {
 			h.logger.Error(LogErrFailedUnsubscribe + (ErrClientNotInitialized).Error())
@@ -272,7 +282,7 @@ func parseSubtopic(subtopic string) (string, error) {
 	return subtopic, nil
 }
 
-func (h *handler) getSubcription(c *session.Client, topics *[]string) ([]Subscription, error) {
+func (h *handler) getSubcriptions(c *session.Client, topics *[]string) ([]Subscription, error) {
 	var subs []Subscription
 	for _, t := range *topics {
 		channelAndSubtopic := channelRegExp.FindStringSubmatch(t)
